pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claweb"
version = "0.1.0"
description = "Intelligent Web Automation Agent based on Tarsier and Vision LLM"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AI-Claw Team", email = "contact@ai-claw.dev"}
]
maintainers = [
    {name = "AI-Claw Team", email = "contact@ai-claw.dev"}
]
keywords = [
    "web-automation",
    "agent",
    "llm",
    "vision",
    "tarsier",
    "playwright",
    "ai"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.40.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "tarsier>=0.4.0",
]

[project.optional-dependencies]
mysql = [
    "mysql-connector-python>=8.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
all = [
    "claweb[mysql,dev]",
]

[project.urls]
Homepage = "https://github.com/ai-claw/claweb"
Documentation = "https://github.com/ai-claw/claweb#readme"
Repository = "https://github.com/ai-claw/claweb"
Issues = "https://github.com/ai-claw/claweb/issues"
Changelog = "https://github.com/ai-claw/claweb/blob/main/CHANGELOG.md"

[project.scripts]
claweb = "claweb.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
claweb = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["claweb"]
skip = [".venv", "venv", "build", "dist"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "build",
    "dist",
    "venv",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.coverage.run]
source = ["src/claweb"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
